backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
easyocr==1.7.2
fastapi==0.116.1
filelock==3.19.1
fsspec==2025.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
joblib==1.5.1
lazy_loader==0.4
lxml==6.0.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
ninja==1.13.0
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
portalocker==3.2.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pydantic==2.11.7
pydantic_core==2.33.2
PyMuPDF==1.26.3
pyparsing==3.2.3
python-bidi==0.6.6
python-dotenv==1.1.1
python-multipart==0.0.20
pywin32==311
PyYAML==6.0.2
qdrant-client==1.15.1
regex==2025.7.34
requests==2.32.5
rsa==4.9.1
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.1.0
setuptools==80.9.0
shapely==2.1.1
sniffio==1.3.1
soupsieve==2.7
starlette==0.47.2
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.6.11
tokenizers==0.21.4
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.55.4
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
